Refactored download helpers into individual files #248
