+++ l2tdevtools/download_helpers/pypi.py
@@ -1,0 +1,125 @@
+# -*- coding: utf-8 -*-
+"""Download helper object implementations."""
+
+from __future__ import unicode_literals
+
+import re
+
+# pylint: disable=wrong-import-position
+import pkg_resources
+
+from l2tdevtools.download_helpers import project
+
+
+class PyPIDownloadHelper(project.ProjectDownloadHelper):
+  """Helps in downloading a PyPI code project."""
+
+  def __init__(self, download_url):
+    """Initializes the download helper.
+
+    Args:
+      download_url (str): download URL.
+
+    Raises:
+      ValueError: if download URL is not supported.
+    """
+    url_segments = download_url.split('/')
+    if (len(url_segments) < 5 or url_segments[2] != 'pypi.python.org' or
+        url_segments[3] != 'pypi'):
+      raise ValueError('Unsupported download URL.')
+
+    super(PyPIDownloadHelper, self).__init__(download_url)
+    self._project_name = url_segments[4]
+
+  def GetLatestVersion(self, unused_project_name, version_definition):
+    """Retrieves the latest version number for a given project name.
+
+    Args:
+      project_name (str): name of the project.
+      version_definition (ProjectVersionDefinition): project version definition
+          or None.
+
+    Returns:
+      str: latest version number or None on error.
+    """
+    if version_definition:
+      earliest_version = version_definition.GetEarliestVersion()
+      if earliest_version and earliest_version[0] == '==':
+        return '.'.join(earliest_version[1:])
+
+    download_url = 'https://pypi.python.org/simple/{0:s}'.format(
+        self._project_name)
+
+    page_content = self.DownloadPageContent(download_url)
+    if not page_content:
+      return
+
+    versions = {}
+    expression_string = (
+        r'../../packages/.*/{0:s}-(?P<version>[\d\.\!]*)'
+        r'\.(tar\.bz2|tar\.gz|zip)#').format(self._project_name)
+    for match in re.finditer(expression_string, page_content):
+      version_string = match.group('version')
+      if version_string:
+        # We need to support PEP440 epoch versioning, which only newly versions
+        # of setuptools support. So this is a bit of hack to handle epochs
+        # but still use setuptools to handle most of the version matching.
+        epoch, _, epoch_version_string = version_string.partition('!')
+        if not epoch_version_string:
+          # Per PEP440, if there's no epoch specified, the epoch is 0.
+          epoch_version_string = epoch
+          epoch = 0
+        else:
+          epoch = int(epoch, 10)
+        version = list(pkg_resources.parse_version(epoch_version_string))
+        # Add the epoch to the version string for comparison.
+        version.insert(0, epoch)
+        version_tuple = (tuple(version), version_string)
+        versions[version_tuple] = version_string
+
+    latest_version = max(versions.keys())
+    if latest_version:
+      # Return the original string that defined the version
+      return versions[latest_version]
+
+  def GetDownloadURL(self, unused_project_name, project_version):
+    """Retrieves the download URL for a given project name and version.
+
+    Args:
+      project_name (str): name of the project.
+      project_version (str): version of the project.
+
+    Returns:
+      str: download URL of the project or None on error.
+    """
+    download_url = (
+        'https://pypi.python.org/pypi/{0:s}/{1!s}').format(
+            self._project_name, project_version)
+
+    page_content = self.DownloadPageContent(download_url)
+    if not page_content:
+      return
+
+    # The format of the project download URL is:
+    # https://pypi.python.org/packages/.*/{project name}-{version}.{extension}
+    expression_string = (
+        '(https://pypi.python.org/packages/.*/'
+        '{0:s}-{1!s}[.](tar[.]bz2|tar[.]gz|zip))').format(
+            self._project_name, project_version)
+    matches = re.findall(expression_string, page_content)
+
+    if matches:
+      return matches[0][0]
+
+    return (
+        'https://pypi.python.org/packages/source/{0:s}/{1:s}/'
+        '{1:s}-{2:s}.tar.gz').format(
+            self._project_name[0], self._project_name, project_version)
+
+  def GetProjectIdentifier(self):
+    """Retrieves the project identifier for a given project name.
+
+    Returns:
+      str: project identifier.
+    """
+    return 'org.python.pypi.{0:s}'.format(self._project_name)
